Start to move into different folders.
- Tests are running.
- Maintainer not yet tested.
- Pylinting not yet done.
